% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute.stats.R
\name{compute.stats}
\alias{compute.stats}
\title{compute the log-likelihood and baseline cumulative hazard}
\usage{
compute.stats(z_hat, D_all, aggregates, betas)
}
\description{
compute the log-likelihood and baseline cumulative hazard
}
