% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dcoxph.R
\name{dcoxph}
\alias{dcoxph}
\title{Run the distributed CoxPH algorithm.}
\usage{
dcoxph(client, expl_vars, time_col, censor_col, serial_col)
}
\description{
Params:
  client: ptmclient::Client instance.
  expl_vars: list of explanatory variables (covariates) to use
  time_col: name of the column that contains the event/censor times
  censor_col: name of the column that explains whether an event occured or
              the patient was censored
  serial_no: set of the serial_no column that explains whether a patient included in the analysis or not
}
\details{
Return:
  data.frame with beta, p-value and confidence interval for each explanatory variable
  and the baseline hazard.
}
